Alzheimer/requirements.txt
streamlit==1.12.0
torch==1.12.1
Pillow==9.1.0
numpy==1.21.0

